+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - feat: Add new plugin for loading routes and mocks in v2 format
 - feat: Pass new method `loadRoutes` to plugins.
 - feat: Add alert when legacy behaviors are loaded.
+- feat: Add option `addRoutesHandlers` to config.
 ### Changed
 - refactor: Move `path` and `watch` options inside files-loader plugin
 - refactor: Reorganize files and folders